Refactoring progetto SHARED
