+++ Shared/Models/ViewModels/PersonaViewModel.cs
@@ -1,0 +1,11 @@
+﻿namespace DemoBlazorApp.Shared.Models.ViewModels
+{
+    public class PersonaViewModel
+    {
+        public int PersonaId { get; set; }
+        public string Cognome { get; set; }
+        public string Nome { get; set; }
+        public string Email { get; set; }
+        public string Telefono { get; set; }
+    }
+}
